refactor(portfolio): remove old thesis card schema, update all references
